Remove onDebug event
Since the debug event is fired for other debuggers too, the Kotlin
language server would launch even in completely unrelated workspaces.

Fix this by activating first when a Kotlin source file has been opened.
The only potential issue with this solution is that no Kotlin debugger
is registered before that, so if the user tries to debug a Kotlin
project before actually opening a file, an error message will pop up.